+++ src/navbar/Navbar.js
@@ -2,6 +2,19 @@ import React from "react";
 import { Link } from "react-router-dom";
 
 export default class Navbar extends React.Component {
+    constructor(props) {
+        super(props);
+        this.state = {
+            side_menu: "",
+        };
+    }
+
+    toggle_side_menu = () => {
+        this.setState({
+            side_menu: this.state.side_menu ? "" : " is-active",
+        });
+    };
+
     render() {
         return (
             <nav
@@ -19,21 +32,26 @@ export default class Navbar extends React.Component {
                         />
                     </a>
 
-                    <a
-                        href="#/"
+                    <span
                         role="button"
-                        className="navbar-burger burger"
+                        className={
+                            "navbar-burger burger" + this.state.side_menu
+                        }
                         aria-label="menu"
                         aria-expanded="false"
                         data-target="side-menu"
+                        onClick={this.toggle_side_menu}
                     >
                         <span aria-hidden="true"></span>
                         <span aria-hidden="true"></span>
                         <span aria-hidden="true"></span>
-                    </a>
+                    </span>
                 </div>
 
-                <div id="side-menu" className="navbar-menu">
+                <div
+                    id="side-menu"
+                    className={"navbar-menu" + this.state.side_menu}
+                >
                     <div className="navbar-start">
                         <div className="navbar-item has-dropdown is-hoverable">
                             <span className="navbar-link">