+++ src/fact/Score.js
@@ -15,19 +15,23 @@ export default class Score extends React.Component {
     }
 
     color_class() {
-        return this.props.score_value < this.state.danger_threshold
-            ? "has-text-danger"
-            : this.props.score_value > this.state.success_threshold
-            ? "has-text-success"
-            : "";
+        return this.props.score_value < this.state.danger_threshold ||
+            this.props.score_value === "<30"
+            ? "has-background-danger"
+            : this.props.score_value > this.state.success_threshold ||
+              this.props.score_value === ">70"
+            ? "has-background-success"
+            : "has-background-warning";
     }
 
     advice() {
-        return this.props.score_value < this.state.danger_threshold
-            ? "Try to get more 'Yes' answers to improve your score."
-            : this.props.score_value > this.state.success_threshold
-            ? "You're doing great!"
-            : "Try to get more 'Yes' answers to improve your score.";
+        return this.props.score_value < this.state.danger_threshold ||
+            this.props.score_value === "<30"
+            ? "Keep going"
+            : this.props.score_value > this.state.success_threshold ||
+              this.props.score_value === ">70"
+            ? "Doing great!"
+            : "Getting there";
     }
 
     render() {
@@ -35,12 +39,33 @@ export default class Score extends React.Component {
             <div className="level-item has-text-centered">
                 <div
                     className="has-tooltip-bottom"
-                    data-tooltip={this.advice()}
+                    data-tooltip={
+                        this.props.label === "legend"
+                            ? "What do the colours mean?"
+                            : this.advice()
+                    }
                 >
                     {this.props.label ? (
-                        <p className="heading">{this.props.label}</p>
-                    ) : null}
-                    <p className={"title " + this.color_class()}>
+                        this.props.label === "legend" ? (
+                            <p className="heading round-score-heading has-text-white">
+                                {this.advice()}
+                            </p>
+                        ) : (
+                            <p className="heading round-score-heading has-text-white">
+                                {this.props.label}
+                            </p>
+                        )
+                    ) : (
+                        <p className="heading round-score-heading has-text-white">
+                            Overall
+                        </p>
+                    )}
+                    <p
+                        className={
+                            "round-score title has-text-white " +
+                            this.color_class()
+                        }
+                    >
                         {this.props.score_value}%
                     </p>
                 </div>