+++ CHANGES.md
@@ -3,7 +3,7 @@
 - [ ] Fix side menu on mobile
 - [ ] Explicit link to questions on the landing page
 - [x] Color code the scores
-- [ ] Add advice based on the scores
+- [x] Add advice based on the scores
 - [ ] Show github badge with score
 - [ ] Report summary of answers
 - [ ] Image map the components to the respective tabs