+++ CHANGES.md
@@ -1,12 +1,12 @@
 0.3.0 (2020-MM-DD)
 ------------------
 - [ ] Fix side menu on mobile
-- [ ] Explicit link to questions on the landing page
+- [x] Explicit link to questions on the landing page
 - [x] Color code the scores
 - [x] Add advice based on the scores
 - [ ] Show github badge with score
 - [ ] Report summary of answers
-- [ ] Image map the components to the respective tabs
+- [x] Image map the components to the respective tabs
 
 0.2.0 (2020-03-01)
 ------------------